requirements.txt
pandas==1.2.2
scikit-learn==0.24.1
scipy==1.6.1
ipython==7.19.0
SQLAlchemy==1.4.17
cx-Oracle==8.2.1
pymssql==2.1.5
psycopg2-binary==2.9.1
openpyxl==3.0.7
textdistance==4.2.1
fastapi==0.65.2
python-slugify==5.0.2
sparse-dot-topn==0.3.1
sqlmodel==0.0.4
python-dotenv==0.19.0
coverage==6.0
requests==2.26.0
uvicorn==0.15.0
pytest==6.2.5
